siteblog/requirements.txt
asgiref==3.8.1
click==7.1.2
click-didyoumean==0.3.1
Django==5.0.7
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
